scripts/compute_positions/requirements.txt
# HORUS Position Pipeline Dependencies

# Core ML
numpy>=1.24.0
umap-learn>=0.5.4
scikit-learn>=1.3.0

# HuggingFace for SAE weights
huggingface_hub>=0.20.0
safetensors>=0.4.0

# HTTP requests for Neuronpedia API
requests>=2.31.0
aiohttp>=3.9.0

# Progress bars
tqdm>=4.66.0

# JSON/compression
orjson>=3.9.0

# Optional: GPU acceleration (uncomment if using NVIDIA GPU)
# cuml-cu12>=24.02
